backend/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
llama-index-core==0.10.27
pinecone-client>=2.2.4
llama-index-vector-stores-pinecone>=0.1.6
llama-index-embeddings-openai>=0.1.5
PyPDF2>=3.0.0,<4.0.0
llama-index-readers-file==0.1.4
pydantic==2.6.4
supabase>=1.0.0,<2.0.0
python-multipart==0.0.9
python-dotenv==1.0.1
tiktoken==0.5.2
PyJWT>=2.6.0
requests>=2.31.0
